Group related modules and refactor commands
