+++ views/base.py
